improvement(interactorstyleunicam): extract a UnicamRotateManipulator
This manipulator the rotation interactor of the Unicam, without the pan and the dolly
