+++ Sources/Interaction/Manipulators/MouseCameraUnicamManipulator/index.js
@@ -1,14 +1,8 @@
-import vtkActor from 'vtk.js/Sources/Rendering/Core/Actor';
 import vtkCompositeCameraManipulator from 'vtk.js/Sources/Interaction/Manipulators/CompositeCameraManipulator';
 import vtkCompositeMouseManipulator from 'vtk.js/Sources/Interaction/Manipulators/CompositeMouseManipulator';
 import vtkInteractorStyleConstants from 'vtk.js/Sources/Rendering/Core/InteractorStyle/Constants';
-import vtkMapper from 'vtk.js/Sources/Rendering/Core/Mapper';
-import vtkOpenGLHardwareSelector from 'vtk.js/Sources/Rendering/OpenGL/HardwareSelector';
-import vtkPicker from 'vtk.js/Sources/Rendering/Core/Picker';
-import vtkSphereSource from 'vtk.js/Sources/Filters/Sources/SphereSource';
+import vtkMouseCameraUnicamRotateManipulator from 'vtk.js/Sources/Interaction/Manipulators/MouseCameraUnicamRotateManipulator';
 
-import { FieldAssociations } from 'vtk.js/Sources/Common/DataModel/DataSet/Constants';
-import { mat4, vec3 } from 'gl-matrix';
 import macro from 'vtk.js/Sources/macro';
 import * as vtkMath from 'vtk.js/Sources/Common/Core/Math';
 
@@ -22,47 +16,16 @@ function vtkMouseCameraUnicamManipulator(publicAPI, model) {
   // Set our className
   model.classHierarchy.push('vtkMouseCameraUnicamManipulator');
 
-  // Setup HardwareSelector and Picker to pick points
-  model.selector = vtkOpenGLHardwareSelector.newInstance({
-    captureZValues: true,
-  });
-  model.selector.setFieldAssociation(
-    FieldAssociations.FIELD_ASSOCIATION_POINTS
-  );
-  model.picker = vtkPicker.newInstance();
-
-  model.isDot = false;
   model.state = States.IS_NONE;
 
-  // Setup focus dot
-  const sphereSource = vtkSphereSource.newInstance();
-  sphereSource.setThetaResolution(6);
-  sphereSource.setPhiResolution(6);
-  const sphereMapper = vtkMapper.newInstance();
-  sphereMapper.setInputConnection(sphereSource.getOutputPort());
-
-  // XXX - Would like to make the focus sphere not be affected by
-  // XXX - the lights-- i.e., always be easily easily seen.  i'm not sure
-  // XXX - how to do that.
-  model.focusSphere = vtkActor.newInstance();
-  model.focusSphere.setMapper(sphereMapper);
-  model.focusSphere.getProperty().setColor(0.89, 0.66, 0.41);
-  model.focusSphere.getProperty().setRepresentationToWireframe();
-
-  //----------------------------------------------------------------------------
-  const updateAndRender = (interactor) => {
-    if (!interactor) {
-      return;
-    }
-
-    if (model.useWorldUpVec) {
-      const camera = interactor.findPokedRenderer().getActiveCamera();
-      if (!vtkMath.areEquals(model.worldUpVec, camera.getViewPlaneNormal())) {
-        camera.setViewUp(model.worldUpVec);
-      }
-    }
-    interactor.render();
-  };
+  model.rotateManipulator = vtkMouseCameraUnicamRotateManipulator.newInstance({
+    button: model.button,
+    shift: model.shift,
+    control: model.control,
+    alt: model.alt,
+    dragEnabled: model.dragEnabled,
+    scrollEnabled: model.scrollEnabled,
+  });
 
   //----------------------------------------------------------------------------
   const normalize = (position, interactor) => {
@@ -119,38 +82,6 @@ function vtkMouseCameraUnicamManipulator(publicAPI, model) {
   };
 
   //----------------------------------------------------------------------------
-  // Rotate the camera by 'angle' degrees about the point <cx, cy, cz>
-  // and around the vector/axis <ax, ay, az>.
-  const rotateCamera = (camera, cx, cy, cz, ax, ay, az, angle) => {
-    const cameraPosition = camera.getPosition();
-    const cameraFocalPoint = camera.getFocalPoint();
-    const cameraViewUp = camera.getViewUp();
-
-    cameraPosition[3] = 1.0;
-    cameraFocalPoint[3] = 1.0;
-    cameraViewUp[3] = 0.0;
-
-    const transform = mat4.create();
-    mat4.identity(transform);
-    mat4.translate(transform, transform, vec3.fromValues(cx, cy, cz));
-    mat4.rotate(transform, transform, angle, vec3.fromValues(ax, ay, az));
-    mat4.translate(transform, transform, vec3.fromValues(-cx, -cy, -cz));
-    const newCameraPosition = [];
-    const newCameraFocalPoint = [];
-    vec3.transformMat4(newCameraPosition, cameraPosition, transform);
-    vec3.transformMat4(newCameraFocalPoint, cameraFocalPoint, transform);
-
-    mat4.identity(transform);
-    mat4.rotate(transform, transform, angle, vec3.fromValues(ax, ay, az));
-    const newCameraViewUp = [];
-    vec3.transformMat4(newCameraViewUp, cameraViewUp, transform);
-
-    camera.setPosition(...newCameraPosition);
-    camera.setFocalPoint(...newCameraFocalPoint);
-    camera.setViewUp(...newCameraViewUp);
-  };
-
-  //----------------------------------------------------------------------------
   const choose = (interactor, position) => {
     const normalizedPosition = normalize(position, interactor);
     const normalizedPreviousPosition = normalize(
@@ -178,110 +109,6 @@ function vtkMouseCameraUnicamManipulator(publicAPI, model) {
       } else {
         model.state = States.IS_DOLLY;
       }
-    }
-  };
-
-  //----------------------------------------------------------------------------
-  const rotate = (interactor, position) => {
-    const renderer = interactor.findPokedRenderer();
-    const normalizedPosition = normalize(position, interactor);
-    const normalizedPreviousPosition = normalize(
-      model.previousPosition,
-      interactor
-    );
-
-    const center = model.focusSphere.getPosition();
-    let normalizedCenter = interactor
-      .getView()
-      .worldToDisplay(...center, renderer);
-    // let normalizedCenter = publicAPI.computeWorldToDisplay(renderer, ...center);
-    normalizedCenter = normalize({ x: center[0], y: center[1] }, interactor);
-    normalizedCenter = [normalizedCenter.x, normalizedCenter.y, center[2]];
-
-    // Squared rad of virtual cylinder
-    const radsq = (1.0 + Math.abs(normalizedCenter[0])) ** 2.0;
-    const op = [normalizedPreviousPosition.x, 0, 0];
-    const oe = [normalizedPosition.x, 0, 0];
-
-    const opsq = op[0] ** 2;
-    const oesq = oe[0] ** 2;
-
-    const lop = opsq > radsq ? 0 : Math.sqrt(radsq - opsq);
-    const loe = oesq > radsq ? 0 : Math.sqrt(radsq - oesq);
-
-    const nop = [op[0], 0, lop];
-    vtkMath.normalize(nop, interactor);
-    const noe = [oe[0], 0, loe];
-    vtkMath.normalize(noe, interactor);
-
-    const dot = vtkMath.dot(nop, noe);
-    if (Math.abs(dot) > 0.0001) {
-      const angle =
-        -2 *
-        Math.acos(vtkMath.clampValue(dot, -1.0, 1.0)) *
-        Math.sign(normalizedPosition.x - normalizedPreviousPosition.x) *
-        publicAPI.getRotationFactor();
-
-      const camera = renderer.getActiveCamera();
-
-      const upVec = model.useWorldUpVec ? model.worldUpVec : camera.getViewUp();
-      vtkMath.normalize(upVec, interactor);
-
-      rotateCamera(camera, ...center, ...upVec, angle);
-
-      const dVec = [];
-      const cameraPosition = camera.getPosition();
-      vtkMath.subtract(cameraPosition, position, dVec);
-
-      let rDist =
-        (normalizedPosition.y - normalizedPreviousPosition.y) *
-        publicAPI.getRotationFactor();
-      vtkMath.normalize(dVec, interactor);
-
-      const atV = camera.getViewPlaneNormal();
-      const upV = camera.getViewUp();
-      const rightV = [];
-      vtkMath.cross(upV, atV, rightV);
-      vtkMath.normalize(rightV, interactor);
-
-      //
-      // The following two tests try to prevent chaotic camera movement
-      // that results from rotating over the poles defined by the
-      // "WorldUpVector".  The problem is the constraint to keep the
-      // camera's up vector in line w/ the WorldUpVector is at odds with
-      // the action of rotating over the top of the virtual sphere used
-      // for rotation.  The solution here is to prevent the user from
-      // rotating the last bit required to "go over the top"-- as a
-      // consequence, you can never look directly down on the poles.
-      //
-      // The "0.99" value is somewhat arbitrary, but seems to produce
-      // reasonable results.  (Theoretically, some sort of clamping
-      // function could probably be used rather than a hard cutoff, but
-      // time constraints prevent figuring that out right now.)
-      //
-      if (model.useWorldUpVec) {
-        const OVER_THE_TOP_THRESHOLD = 0.99;
-        if (vtkMath.dot(upVec, atV) > OVER_THE_TOP_THRESHOLD && rDist < 0) {
-          rDist = 0;
-        }
-        if (vtkMath.dot(upVec, atV) < -OVER_THE_TOP_THRESHOLD && rDist > 0) {
-          rDist = 0;
-        }
-      }
-
-      rotateCamera(camera, ...center, ...rightV, rDist);
-
-      if (
-        model.useWorldUpVec &&
-        !vtkMath.areEquals(upVec, camera.getViewPlaneNormal())
-      ) {
-        camera.setViewUp(...upVec);
-      }
-
-      model.previousPosition = position;
-
-      renderer.resetCameraClippingRange();
-      updateAndRender(interactor);
     }
   };
 
@@ -315,7 +142,7 @@ function vtkMouseCameraUnicamManipulator(publicAPI, model) {
     camera.translate(...offset);
 
     renderer.resetCameraClippingRange();
-    updateAndRender(interactor);
+    interactor.render();
   };
 
   //----------------------------------------------------------------------------
@@ -354,7 +181,7 @@ function vtkMouseCameraUnicamManipulator(publicAPI, model) {
     camera.translate(...offset2);
 
     renderer.resetCameraClippingRange();
-    updateAndRender(interactor);
+    interactor.render();
   };
 
   //----------------------------------------------------------------------------
@@ -367,39 +194,10 @@ function vtkMouseCameraUnicamManipulator(publicAPI, model) {
     model.time = Date.now() / 1000.0;
     model.dist = 0;
 
-    // Finds the point under the cursor.
-    // Note: If no object has been rendered to the pixel (X, Y), then
-    // vtkPicker will return a z-value with depth equal
-    // to the distance from the camera's position to the focal point.
-    // This seems like an arbitrary, but perhaps reasonable, default value.
-    model.selector.attach(interactor.getView(), renderer);
+    // Picking is delegated to the rotate manipulator
+    model.rotateManipulator.onButtonDown(interactor, renderer, position);
 
-    model.selector.setArea(position.x, position.y, position.x, position.y);
-
-    const selections = model.selector.select();
-    if (selections && selections.length !== 0) {
-      model.downPoint = selections[0].getProperties().worldPosition;
-    } else {
-      model.picker.pick([position.x, position.y, position.z], renderer);
-      model.downPoint = model.picker.getPickPosition();
-    }
-
-    const normalizedPosition = normalize(position, interactor);
-    // borderRatio defines the percentage of the screen size that is considered to be
-    // the border of the screen on each side
-    const borderRatio = 0.1;
-    // If someone has already clicked to make a dot and they're not clicking
-    // on it now, OR if the user is clicking on the perimeter of the screen,
-    // then we want to go into rotation mode.
-    if (
-      Math.abs(normalizedPosition.x) > 1 - borderRatio ||
-      Math.abs(normalizedPosition.y) > 1 - borderRatio ||
-      model.isDot
-    ) {
-      model.state = States.IS_ROTATE;
-    } else {
-      model.state = States.IS_NONE;
-    }
+    model.downPoint = model.rotateManipulator.getDownPoint();
   };
 
   //----------------------------------------------------------------------------
@@ -408,25 +206,22 @@ function vtkMouseCameraUnicamManipulator(publicAPI, model) {
       return;
     }
 
-    switch (model.state) {
-      case States.IS_NONE:
-        choose(interactor, position);
-        // publicAPI.invokeInteractionEvent({ type: 'InteractionEvent' });
-        break;
-      case States.IS_ROTATE:
-        rotate(interactor, position);
-        // publicAPI.invokeInteractionEvent({ type: 'InteractionEvent' });
-        break;
-      case States.IS_PAN:
-        pan(interactor, position);
-        // publicAPI.invokeInteractionEvent({ type: 'InteractionEvent' });
-        break;
-      case States.IS_DOLLY:
-        dolly(interactor, position);
-        // publicAPI.invokeInteractionEvent({ type: 'InteractionEvent' });
-        break;
-      default:
-        break;
+    if (model.rotateManipulator.getState() === States.IS_ROTATE) {
+      model.rotateManipulator.onMouseMove(interactor, renderer, position);
+    } else {
+      switch (model.state) {
+        case States.IS_NONE:
+          choose(interactor, position);
+          break;
+        case States.IS_PAN:
+          pan(interactor, position);
+          break;
+        case States.IS_DOLLY:
+          dolly(interactor, position);
+          break;
+        default:
+          break;
+      }
     }
 
     model.previousPosition = position;
@@ -434,44 +229,42 @@ function vtkMouseCameraUnicamManipulator(publicAPI, model) {
 
   //--------------------------------------------------------------------------
   publicAPI.onButtonUp = (interactor) => {
-    const renderer = interactor.findPokedRenderer();
     model.buttonPressed = false;
-
-    if (model.state === States.IS_ROTATE && model.isDot) {
-      renderer.removeActor(model.focusSphere);
-      model.isDot = false;
-    } else if (model.state === States.IS_NONE) {
-      if (model.isDot) {
-        renderer.removeActor(model.focusSphere);
-        model.isDot = false;
-      } else {
-        model.focusSphere.setPosition(...model.downPoint);
-
-        const camera = renderer.getActiveCamera();
-        const cameraPosition = camera.getPosition();
-        const cameraToPointVec = [];
-
-        vtkMath.subtract(model.downPoint, cameraPosition, cameraToPointVec);
-        if (camera.getParallelProjection()) {
-          vtkMath.multiplyScalar(cameraToPointVec, camera.getParallelScale());
-        }
-
-        const atV = camera.getDirectionOfProjection();
-        vtkMath.normalize(atV, interactor);
-
-        // Scales the focus dot so it always appears the same size
-        const scale = 0.02 * vtkMath.dot(atV, cameraToPointVec);
-
-        model.focusSphere.setScale(scale, scale, scale);
-
-        renderer.addActor(model.focusSphere);
-
-        model.isDot = true;
-      }
+    if (model.state === States.IS_NONE) {
+      model.rotateManipulator.onButtonUp(interactor);
     }
+    model.state = States.IS_NONE;
+  };
 
-    renderer.resetCameraClippingRange();
-    updateAndRender(interactor);
+  publicAPI.getUseWorldUpVec = () => {
+    return model.rotateManipulator.getUseWorldUpVec();
+  };
+  publicAPI.setUseWorldUpVec = (useWorldUpVec) => {
+    model.rotateManipulator.setUseWorldUpVec(useWorldUpVec);
+  };
+  publicAPI.getWorldUpVec = () => {
+    return model.rotateManipulator.getWorldUpVec();
+  };
+  publicAPI.setWorldUpVec = (x, y, z) => {
+    model.rotateManipulator.setWorldUpVec(x, y, z);
+  };
+  publicAPI.getUseHardwareSelector = () => {
+    return model.rotateManipulator.getUseHardwareSelector();
+  };
+  publicAPI.setUseHardwareSelector = (useHardwareSelector) => {
+    model.rotateManipulator.setUseHardwareSelector(useHardwareSelector);
+  };
+  publicAPI.getFocusSphereColor = () => {
+    model.rotateManipulator.getFocusSphereColor();
+  };
+  publicAPI.setFocusSphereColor = (r, g, b) => {
+    model.rotateManipulator.setFocusSphereColor(r, g, b);
+  };
+  publicAPI.getFocusSphereRadiusFactor = () => {
+    return model.rotateManipulator.getFocusSphereRadiusFactor();
+  };
+  publicAPI.setFocusSphereRadiusFactor = (focusSphereRadiusFactor) => {
+    model.rotateManipulator.setFocusSphereRadiusFactor(focusSphereRadiusFactor);
   };
 }
 
@@ -479,11 +272,7 @@ function vtkMouseCameraUnicamManipulator(publicAPI, model) {
 // Object factory
 // ----------------------------------------------------------------------------
 
-const DEFAULT_VALUES = {
-  useWorldUpVec: true,
-  // set WorldUpVector to be z-axis by default
-  worldUpVec: [0, 0, 1],
-};
+const DEFAULT_VALUES = {};
 
 // ----------------------------------------------------------------------------
 
@@ -494,10 +283,6 @@ export function extend(publicAPI, model, initialValues = {}) {
   macro.obj(publicAPI, model);
   vtkCompositeCameraManipulator.extend(publicAPI, model, initialValues);
   vtkCompositeMouseManipulator.extend(publicAPI, model, initialValues);
-
-  // Create get-set macros
-  macro.setGet(publicAPI, model, ['useWorldUpVec']);
-  macro.setGetArray(publicAPI, model, ['worldUpVec'], 3);
 
   // Object specific methods
   vtkMouseCameraUnicamManipulator(publicAPI, model);
