+++ Sources/Interaction/Manipulators/index.js
@@ -8,6 +8,7 @@ import vtkKeyboardCameraManipulator from './KeyboardCameraManipulator';
 import vtkMouseCameraAxisRotateManipulator from './MouseCameraAxisRotateManipulator';
 import vtkMouseCameraSliceManipulator from './MouseCameraSliceManipulator';
 import vtkMouseCameraUnicamManipulator from './MouseCameraUnicamManipulator';
+import vtkMouseCameraUnicamRotateManipulator from './MouseCameraUnicamRotateManipulator';
 import vtkMouseCameraTrackballFirstPersonManipulator from './MouseCameraTrackballFirstPersonManipulator';
 import vtkMouseCameraTrackballMultiRotateManipulator from './MouseCameraTrackballMultiRotateManipulator';
 import vtkMouseCameraTrackballPanManipulator from './MouseCameraTrackballPanManipulator';
@@ -29,6 +30,7 @@ export default {
   vtkMouseCameraAxisRotateManipulator,
   vtkMouseCameraSliceManipulator,
   vtkMouseCameraUnicamManipulator,
+  vtkMouseCameraUnicamRotateManipulator,
   vtkMouseCameraTrackballFirstPersonManipulator,
   vtkMouseCameraTrackballMultiRotateManipulator,
   vtkMouseCameraTrackballPanManipulator,
