+++ Sources/Interaction/Style/InteractorStyleUnicam/index.js
@@ -27,6 +27,24 @@ function vtkInteractorStyleUnicam(publicAPI, model) {
   publicAPI.setWorldUpVec = (x, y, z) => {
     model.unicamManipulator.setWorldUpVec(x, y, z);
   };
+  publicAPI.getUseHardwareSelector = () => {
+    return model.unicamManipulator.getUseHardwareSelector();
+  };
+  publicAPI.setUseHardwareSelector = (useHardwareSelector) => {
+    model.unicamManipulator.setUseHardwareSelector(useHardwareSelector);
+  };
+  publicAPI.getFocusSphereColor = () => {
+    model.unicamManipulator.getFocusSphereColor();
+  };
+  publicAPI.setFocusSphereColor = (r, g, b) => {
+    model.unicamManipulator.setFocusSphereColor(r, g, b);
+  };
+  publicAPI.getFocusSphereRadiusFactor = () => {
+    return model.unicamManipulator.getFocusSphereRadiusFactor();
+  };
+  publicAPI.setFocusSphereRadiusFactor = (focusSphereRadiusFactor) => {
+    model.unicamManipulator.setFocusSphereRadiusFactor(focusSphereRadiusFactor);
+  };
 }
 
 // ----------------------------------------------------------------------------
